+++ public/dialog_font_size.js
@@ -1,12 +1,12 @@
 import {ChangeEditorOptionEvent} from '/change_editor_option_event.js';
 import {DialogPrompt} from '/dialog_prompt.js';
 import {Res} from '/res.js';
+import {SaveOptionsEditorEvent} from '/save_options_editor_event.js';
 
 export class DialogFontSize extends DialogPrompt {
 
-    constructor(core, initialValue) {
+    constructor(initialValue) {
         super();
-        this.core = core;
         this.initialValue = initialValue.replace('px', '');
     }
 
@@ -26,7 +26,7 @@ export class DialogFontSize extends DialogPrompt {
 
     submit() {
         super.submit();
-        this.core.saveRendererOptions();
+        document.dispatchEvent(new SaveOptionsEditorEvent({renderer: true}));
     }
 
     reset() {
@@ -34,6 +34,5 @@ export class DialogFontSize extends DialogPrompt {
         document.dispatchEvent(
             new ChangeEditorOptionEvent('fontSize', this.initialValue + 'px')
         );
-        this.core.saveRendererOptions();
     }
 }