+++ public/dialog_scroll_past_end.js
@@ -1,12 +1,12 @@
 import {ChangeEditorOptionEvent} from '/change_editor_option_event.js';
 import {DialogSelect} from '/dialog_select.js';
 import {Res} from '/res.js';
+import {SaveOptionsEditorEvent} from '/save_options_editor_event.js';
 
 export class DialogScrollPastEnd extends DialogSelect {
 
-    constructor(core, initialValue) {
+    constructor(initialValue) {
         super();
-        this.core = core;
         this.initialValue = initialValue;
     }
 
@@ -23,7 +23,7 @@ export class DialogScrollPastEnd extends DialogSelect {
 
     submit() {
         super.submit();
-        this.core.saveRendererOptions();
+        document.dispatchEvent(new SaveOptionsEditorEvent({renderer: true}));
     }
 
     reset() {
@@ -31,6 +31,5 @@ export class DialogScrollPastEnd extends DialogSelect {
         document.dispatchEvent(
             new ChangeEditorOptionEvent('scrollPastEnd', this.initialValue)
         );
-        this.core.saveRendererOptions();
     }
 }