+++ public/dialog_theme.js
@@ -2,12 +2,12 @@ import {ChangeEditorOptionEvent} from '/change_editor_option_event.js';
 import {DialogSelect} from '/dialog_select.js';
 import {MaterialHelper} from '/material_helper.js';
 import {Res} from '/res.js';
+import {SaveOptionsEditorEvent} from '/save_options_editor_event.js';
 
 export class DialogTheme extends DialogSelect {
 
-    constructor(core, initialValue) {
+    constructor(initialValue) {
         super();
-        this.core = core;
         this.initialValue = initialValue;
     }
 
@@ -55,7 +55,7 @@ export class DialogTheme extends DialogSelect {
 
     submit() {
         super.submit();
-        this.core.saveRendererOptions();
+        document.dispatchEvent(new SaveOptionsEditorEvent({renderer: true}));
     }
 
     reset() {
@@ -63,6 +63,5 @@ export class DialogTheme extends DialogSelect {
         document.dispatchEvent(
             new ChangeEditorOptionEvent('theme', this.initialValue)
         );
-        this.core.saveRendererOptions();
     }
 }